Update attributecert.go
Introduce three proof of concept.

a) show that we can read platform cert as asn1 structure map and convert back to new Der certificate file. Check the original and rebuild cert are identical.

b) show that we can read platform cert and replace signature and rebuild the certficate.

c)  show that we can read a ekCert and extract issuer and serial number. Create a attriguteCert structure and populate holder field with this data. Show that we are able to correctly parse RDN sequence from attributeCert data. Since we can build certficate from attributeCert structure, this shows we can add field value .